feat: more useful staking events in `Governance`
Previously, the details of V1 rewards were emitted by `UserProxy`.
This made it harder to query all the information pertaining to the
V1 staking that goes through V2, because each user has their own
`UserProxy` address.

Plus, the event `WithdrawLQTY` suffered from an issue which has
already been fixed for the corresponding event in `UserProxy`,
which is that a user can attempt to withdraw more than their V1
staking LQTY balance, in which case the withdrawal gets truncated,
but the parameter `withdrawnLQTY` didn't reflect this truncation.

As we don't really need to have duplicate events (`UserProxy` &
`Governance`) anyway, we eliminate them from `UserProxy` but
expose the details to `Governance`, so that it may emit more
useful events of its own.